requirements.txt
streamlit>=1.37.0
google-generativeai>=0.8.0
PyPDF2>=3.0.0
